web/src/requirements.txt
Flask==1.1.2
Werkzeug==1.0.1
google-auth
google-cloud-storage
google-cloud-secret-manager
python-dotenv